# refactor(sumologicexporter): use golang.org/x/exp/slices for sorting fields