--- a/pkg/exporter/sumologicexporter/go.mod
+++ b/pkg/exporter/sumologicexporter/go.mod
@@ -12,6 +12,7 @@ require (
 	go.opentelemetry.io/collector/model v0.47.0
 	go.uber.org/multierr v1.8.0
 	go.uber.org/zap v1.21.0
+	golang.org/x/exp v0.0.0-20220328175248-053ad81199eb
 )
 
 require (