This PR uses generic slice sorting algorithm and saves yet another allocation in `fields.string()`.

Related article: https://eli.thegreenplace.net/2022/faster-sorting-with-go-generics/

```
benchcmp old.txt new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark              old ns/op     new ns/op     delta
BenchmarkFields-16     646           618           -4.32%
BenchmarkFields-16     642           603           -6.03%
BenchmarkFields-16     636           614           -3.35%
BenchmarkFields-16     640           616           -3.72%
BenchmarkFields-16     648           612           -5.56%

benchmark              old allocs     new allocs     delta
BenchmarkFields-16     11             10             -9.09%
BenchmarkFields-16     11             10             -9.09%
BenchmarkFields-16     11             10             -9.09%
BenchmarkFields-16     11             10             -9.09%
BenchmarkFields-16     11             10             -9.09%

benchmark              old bytes     new bytes     delta
BenchmarkFields-16     368           344           -6.52%
BenchmarkFields-16     368           344           -6.52%
BenchmarkFields-16     368           344           -6.52%
BenchmarkFields-16     368           344           -6.52%
BenchmarkFields-16     368           344           -6.52%
```

**EDIT**: the linter fails because we don't have full generics support in golangci-lint just yet: https://github.com/golangci/golangci-lint/issues/2649