@swiatekm-sumo 2022-04-05
I'm not sure if I'm reading this right, but from https://github.com/golangci/golangci-lint/issues/2649 it sounds like the linter should work if we explicitly set go version to 1.18 in the config? The issue description claims that they automatically disable linters which can't handle generics for go 1.18. @pmalek-sumo ?